requirements.txt
attrs==19.3.0
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
elasticsearch==7.5.1
idna==2.8
pystalk==0.6.0
pyTelegramBotAPI==3.6.7
python-dateutil==2.8.1
PyYAML==5.3
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
urllib3==1.25.8
